unified action interfaces between collections and cores
