+++ src/QueryType/Server/Collections/Query/Action/Delete.php
@@ -4,7 +4,6 @@ namespace Solarium\QueryType\Server\Collections\Query\Action;
 
 use Solarium\QueryType\Server\Collections\Query\Query as CollectionsQuery;
 use Solarium\QueryType\Server\Collections\Result\DeleteResult;
-use Solarium\QueryType\Server\Query\Action\AbstractAsyncAction;
 
 /**
  * Class Delete.
